pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "langgraph_checkpoint_snowflake"
version = "0.1.0"
description = "Snowflake checkpoint saver implementation for LangGraph"
authors = [{name = "Kamal", email = "skamalj@github.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain-core",
    "langgraph",
    "snowflake-connector-python"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "black",
    "isort",
    "mypy"
]

[tool.hatch.build.targets.wheel]
packages = ["src/langgraph_checkpoint_snowflake"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3